% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cali_prop99}
\alias{cali_prop99}
\title{Per capita cigarette consumption data by state/year and treatment by California's Tobacco Control Program.
Used as an example of synthetic control method.}
\format{
A tibble with 1,209 rows and 4 columns:
\describe{
\item{state}{State in the US.}
\item{year}{Year.}
\item{packs_per_capita}{Total annual cigarette packs per capita.}
\item{treated}{1 for treated by California's Tobacco Control Program, 0 otherwise.}
}
}
\source{
"Synthetic Control Methods for Comparative Case Studies: Estimating the Effect of California's Tobacco Control Program" by Abadie, Diamond, and Hainmueller
}
\usage{
cali_prop99
}
\description{
Per capita cigarette consumption data by state/year and treatment by California's Tobacco Control Program.
Used as an example of synthetic control method.
}
\keyword{datasets}
