% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicecon-package.R
\docType{package}
\name{classicecon-package}
\alias{classicecon}
\alias{classicecon-package}
\title{classicecon: Classic Economics Datasets}
\description{
Some datasets that I use in class to demonstrate empirical results from classic economics papers.
}
\author{
\strong{Maintainer}: Andrew Griffen \email{andrew.griffen@gmail.com}

}
\keyword{internal}
