% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw}
\alias{nsw}
\title{National Supported Work data for males.}
\format{
A tibble with 722 rows and 12 columns:
\describe{
\item{treated}{Indicator variable for treated.}
\item{age}{Age in years.}
\item{educ}{Education in years.}
\item{black}{Indicator variable for black.}
\item{hispanic}{Indicator variable for Hispanic.}
\item{married}{Indicator variable for married.}
\item{nodegree}{Indicator variable for no degree.}
\item{re74}{Earnings in 1974.}
\item{re75}{Earnings in 1975.}
\item{re78}{Earnings in 1978.}
\item{data}{Indicator for early random assignment.}
\item{data}{Dataset.}
}
}
\source{
LaLonde (1986).
}
\usage{
nsw
}
\description{
National Supported Work data for males.
}
\keyword{datasets}
