% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psid}
\alias{psid}
\title{PSID comparison data.}
\format{
A tibble with 722 rows and 12 columns:
\describe{
\item{age}{Age in years.}
\item{educ}{Education in years.}
\item{black}{Indicator variable for black.}
\item{hispanic}{Indicator variable for Hispanic.}
\item{married}{Indicator variable for married.}
\item{nodegree}{Indicator variable for no degree.}
\item{re74}{Earnings in 1974.}
\item{re75}{Earnings in 1975.}
\item{re79}{Earnings in 1978.}
\item{data}{Dataset.}
}
}
\source{
Panel Study of Income Dynamics.
}
\usage{
psid
}
\description{
PSID comparison data.
}
\keyword{datasets}
