% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Survival data from the Titanic disaster.}
\format{
A tibble with 1,309 rows and 7 columns:
\describe{
\item{survived}{Survived.}
\item{passenger_class}{Passenger class.}
\item{sex}{Male or Female.}
\item{age}{Age in years.}
\item{sibling_spouse}{Number of siblings / spouse on board.}
\item{parents_children}{Number of parents / children on board.}
\item{name}{Name of passenger.}
}
}
\usage{
titanic
}
\description{
Survival data from the Titanic disaster.
}
\keyword{datasets}
